Split binding resolution API tests
Move shared binding-resolution setup helpers into a support module and
separate brief endpoint coverage from resolved-bindings endpoint tests.
Remove the file-level `too-many-lines` suppression from the original module.
